Refactor pb call
