+++ state/client.go
@@ -3,15 +3,50 @@ package state
 import (
 	"context"
 	"fmt"
+	pbDiscovery "github.com/cloudquery/plugin-pb-go/pb/discovery/v1"
+	stateV3 "github.com/cloudquery/plugin-sdk/v4/internal/clients/state/v3"
 	"github.com/cloudquery/plugin-sdk/v4/plugin"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
+	"slices"
 )
 
 const defaultMaxMsgSizeInBytes = 100 * 1024 * 1024 // 100 MiB
 
+type Client interface {
+	SetKey(ctx context.Context, key string, value string) error
+	GetKey(ctx context.Context, key string) (string, error)
+	Flush(ctx context.Context) error
+	Close() error
+}
+
+type ClientOptions struct {
+	Versioned bool
+}
+
 type ConnectionOptions struct {
 	MaxMsgSizeInBytes int
+}
+
+// NewClient returns a client object on which you must call Close().
+func NewClient(ctx context.Context, conn *grpc.ClientConn, tableName string) (Client, error) {
+	return NewClientWithOptions(ctx, conn, tableName, ClientOptions{})
+}
+
+// NewClientWithOptions returns a client object on which you must call Close().
+func NewClientWithOptions(ctx context.Context, conn *grpc.ClientConn, tableName string, opts ClientOptions) (Client, error) {
+	discoveryClient := pbDiscovery.NewDiscoveryClient(conn)
+	versions, err := discoveryClient.GetVersions(ctx, &pbDiscovery.GetVersions_Request{})
+	if err != nil {
+		return nil, err
+	}
+	if slices.Contains(versions.Versions, 3) {
+		if opts.Versioned {
+			return stateV3.NewClientWithTable(ctx, conn, stateV3.VersionedTable(tableName))
+		}
+		return stateV3.NewClient(ctx, conn, tableName)
+	}
+	return nil, fmt.Errorf("please upgrade your state backend plugin. state supporting version 3 plugin has %v", versions.Versions)
 }
 
 // NewGrpcConnectedClient returns a state client and initialises the gRPC connection to the state backend with a 100MiB max message size.
@@ -48,3 +83,24 @@ func NewGrpcConnectedClientWithOptions(ctx context.Context, backendOpts *plugin.
 
 	return stateClient, nil
 }
+
+type NoOpClient struct{}
+
+func (*NoOpClient) SetKey(_ context.Context, _ string, _ string) error {
+	return nil
+}
+
+func (*NoOpClient) GetKey(_ context.Context, _ string) (string, error) {
+	return "", nil
+}
+
+func (*NoOpClient) Flush(_ context.Context) error {
+	return nil
+}
+
+func (*NoOpClient) Close() error {
+	return nil
+}
+
+// static check
+var _ Client = (*NoOpClient)(nil)
